refactor: consolidate internal widgets into index file
