+++ apps/tissuumaps/src/components/controls/MarkerConfigWidget/index.tsx
@@ -1,6 +1,9 @@
-import { ConstantMarkerConfigWidget } from "./ConstantMarkerConfigWidget";
-import { FromMarkerConfigWidget } from "./FromMarkerConfigWidget";
-import { GroupByMarkerConfigWidget } from "./GroupByMarkerConfigWidget";
+import { useTableColumnSelector } from "../../../hooks/useTableColumnSelector";
+import { useTissUUmaps } from "../../../store";
+import { Field, FieldLabel } from "../../common/field";
+import { SimpleAsyncCombobox } from "../../common/simple-combobox";
+import { SimpleSelect } from "../../common/simple-select";
+import { markers } from "./markers";
 import { type MarkerConfigWidgetState } from "./useMarkerConfigWidget";
 
 export { ActiveMarkerConfigValue } from "./ActiveMarkerConfigValue";
@@ -23,4 +26,151 @@ export function MarkerConfigWidget({
     case "groupBy":
       return <GroupByMarkerConfigWidget state={state} className={className} />;
   }
+}
+
+type ConstantMarkerConfigWidgetProps = {
+  state: MarkerConfigWidgetState;
+  className?: string;
+};
+
+function ConstantMarkerConfigWidget({
+  state,
+  className,
+}: ConstantMarkerConfigWidgetProps) {
+  const { currentConstantValue: value, setCurrentConstantValue: setValue } =
+    state;
+
+  return (
+    <div className={className}>
+      <Field>
+        <FieldLabel>Marker</FieldLabel>
+        <SimpleSelect
+          items={markers}
+          itemLabel={(marker) => (
+            <>
+              {marker.icon} {marker.label}
+            </>
+          )}
+          itemValue={(marker) => marker.value}
+          value={value}
+          onValueChange={(value) => {
+            if (value !== null) {
+              setValue(value);
+            }
+          }}
+        />
+      </Field>
+    </div>
+  );
+}
+
+type FromMarkerConfigWidgetProps = {
+  state: MarkerConfigWidgetState;
+  className?: string;
+};
+
+function FromMarkerConfigWidget({
+  state,
+  className,
+}: FromMarkerConfigWidgetProps) {
+  const {
+    currentFromTable: table,
+    currentFromColumn: column,
+    setCurrentFromTable: setTable,
+    setCurrentFromColumn: setColumn,
+  } = state;
+
+  const tables = useTissUUmaps((state) => state.tables);
+
+  const { suggestTableColumnQueries, resolveTableColumnQuery } =
+    useTableColumnSelector(table);
+
+  return (
+    <div className={className}>
+      <div className="grid grid-cols-2 gap-x-2">
+        <Field>
+          <FieldLabel>Source table</FieldLabel>
+          <SimpleSelect
+            items={tables}
+            itemLabel={(table) => table.name}
+            itemValue={(table) => table.id}
+            value={table}
+            onValueChange={setTable}
+          />
+        </Field>
+        <Field disabled={table === null}>
+          <FieldLabel>Source column</FieldLabel>
+          <SimpleAsyncCombobox
+            suggestQueries={suggestTableColumnQueries}
+            getItem={resolveTableColumnQuery}
+            itemQuery={(column) => column}
+            selectedItem={column}
+            onSelectedItemChange={setColumn}
+          />
+        </Field>
+      </div>
+    </div>
+  );
+}
+
+type GroupByMarkerConfigWidgetProps = {
+  state: MarkerConfigWidgetState;
+  className?: string;
+};
+
+function GroupByMarkerConfigWidget({
+  state,
+  className,
+}: GroupByMarkerConfigWidgetProps) {
+  const {
+    currentGroupByTable: table,
+    currentGroupByColumn: column,
+    currentGroupByMap: map,
+    setCurrentGroupByTable: setTable,
+    setCurrentGroupByColumn: setColumn,
+    setCurrentGroupByMap: setMap,
+  } = state;
+
+  const tables = useTissUUmaps((state) => state.tables);
+  const markerMaps = useTissUUmaps((state) => state.markerMaps);
+
+  const { suggestTableColumnQueries, resolveTableColumnQuery } =
+    useTableColumnSelector(table);
+
+  return (
+    <div className={className}>
+      <div className="grid grid-cols-2 gap-x-2">
+        <Field>
+          <FieldLabel>Source table</FieldLabel>
+          <SimpleSelect
+            items={tables}
+            itemLabel={(table) => table.name}
+            itemValue={(table) => table.id}
+            value={table}
+            onValueChange={setTable}
+          />
+        </Field>
+        <Field disabled={table === null}>
+          <FieldLabel>Source column</FieldLabel>
+          <SimpleAsyncCombobox
+            suggestQueries={suggestTableColumnQueries}
+            getItem={resolveTableColumnQuery}
+            itemQuery={(column) => column}
+            selectedItem={column}
+            onSelectedItemChange={setColumn}
+          />
+        </Field>
+      </div>
+      <Field>
+        <FieldLabel>Marker map</FieldLabel>
+        <SimpleSelect
+          items={markerMaps}
+          itemLabel={(markerMap) => markerMap.name}
+          itemValue={(markerMap) => markerMap.id}
+          value={map}
+          onValueChange={setMap}
+        />
+      </Field>
+    </div>
+  );
 }
