--- apps/tissuumaps/src/components/controls/OpacityConfigWidget/FromOpacityConfigWidget.tsx
@@ -1,55 +1,0 @@
-import { useTableColumnSelector } from "../../../hooks/useTableColumnSelector";
-import { useTissUUmaps } from "../../../store";
-import { Field, FieldLabel } from "../../common/field";
-import { SimpleAsyncCombobox } from "../../common/simple-combobox";
-import { SimpleSelect } from "../../common/simple-select";
-import { type OpacityConfigWidgetState } from "./useOpacityConfigWidget";
-
-export type FromOpacityConfigWidgetProps = {
-  state: OpacityConfigWidgetState;
-  className?: string;
-};
-
-export function FromOpacityConfigWidget({
-  state,
-  className,
-}: FromOpacityConfigWidgetProps) {
-  const {
-    currentFromTable: table,
-    currentFromColumn: column,
-    setCurrentFromTable: setTable,
-    setCurrentFromColumn: setColumn,
-  } = state;
-
-  const tables = useTissUUmaps((state) => state.tables);
-
-  const { suggestTableColumnQueries, resolveTableColumnQuery } =
-    useTableColumnSelector(table);
-
-  return (
-    <div className={className}>
-      <div className="grid grid-cols-2 gap-x-2">
-        <Field>
-          <FieldLabel>Source table</FieldLabel>
-          <SimpleSelect
-            items={tables}
-            itemLabel={(table) => table.name}
-            itemValue={(table) => table.id}
-            value={table}
-            onValueChange={setTable}
-          />
-        </Field>
-        <Field disabled={table === null}>
-          <FieldLabel>Source column</FieldLabel>
-          <SimpleAsyncCombobox
-            suggestQueries={suggestTableColumnQueries}
-            getItem={resolveTableColumnQuery}
-            itemQuery={(column) => column}
-            selectedItem={column}
-            onSelectedItemChange={setColumn}
-          />
-        </Field>
-      </div>
-    </div>
-  );
-}
