--- apps/tissuumaps/src/components/controls/OpacityConfigWidget/GroupByOpacityConfigWidget.tsx
@@ -1,68 +1,0 @@
-import { useTableColumnSelector } from "../../../hooks/useTableColumnSelector";
-import { useTissUUmaps } from "../../../store";
-import { Field, FieldLabel } from "../../common/field";
-import { SimpleAsyncCombobox } from "../../common/simple-combobox";
-import { SimpleSelect } from "../../common/simple-select";
-import { type OpacityConfigWidgetState } from "./useOpacityConfigWidget";
-
-export type GroupByOpacityConfigWidgetProps = {
-  state: OpacityConfigWidgetState;
-  className?: string;
-};
-
-export function GroupByOpacityConfigWidget({
-  state,
-  className,
-}: GroupByOpacityConfigWidgetProps) {
-  const {
-    currentGroupByTable: table,
-    currentGroupByColumn: column,
-    currentGroupByMap: map,
-    setCurrentGroupByTable: setTable,
-    setCurrentGroupByColumn: setColumn,
-    setCurrentGroupByMap: setMap,
-  } = state;
-
-  const tables = useTissUUmaps((state) => state.tables);
-  const opacityMaps = useTissUUmaps((state) => state.opacityMaps);
-
-  const { suggestTableColumnQueries, resolveTableColumnQuery } =
-    useTableColumnSelector(table);
-
-  return (
-    <div className={className}>
-      <div className="grid grid-cols-2 gap-x-2">
-        <Field>
-          <FieldLabel>Source table</FieldLabel>
-          <SimpleSelect
-            items={tables}
-            itemLabel={(table) => table.name}
-            itemValue={(table) => table.id}
-            value={table}
-            onValueChange={setTable}
-          />
-        </Field>
-        <Field disabled={table === null}>
-          <FieldLabel>Source column</FieldLabel>
-          <SimpleAsyncCombobox
-            suggestQueries={suggestTableColumnQueries}
-            getItem={resolveTableColumnQuery}
-            itemQuery={(column) => column}
-            selectedItem={column}
-            onSelectedItemChange={setColumn}
-          />
-        </Field>
-      </div>
-      <Field>
-        <FieldLabel>Opacity map</FieldLabel>
-        <SimpleSelect
-          items={opacityMaps}
-          itemLabel={(opacityMap) => opacityMap.name}
-          itemValue={(opacityMap) => opacityMap.id}
-          value={map}
-          onValueChange={setMap}
-        />
-      </Field>
-    </div>
-  );
-}
